extracted only the changes in logic from feature/laion/haproxy-thread branch
