+++ lethean-wallet-gui.pro
@@ -39,7 +39,8 @@ HEADERS += \
     src/libwalletqt/UnsignedTransaction.h \
     MainApp.h \
     src/libwalletqt/Haproxy.h \
-    src/libwalletqt/HTTPResponse.h
+    src/libwalletqt/HTTPResponse.h \
+    src/libwalletqt/Thread.h
 
 SOURCES += main.cpp \
     filter.cpp \
@@ -64,7 +65,8 @@ SOURCES += main.cpp \
     src/libwalletqt/UnsignedTransaction.cpp \
     MainApp.cpp \
     src/libwalletqt/Haproxy.cpp \
-    src/libwalletqt/HTTPResponse.cpp
+    src/libwalletqt/HTTPResponse.cpp \
+    src/libwalletqt/Thread.cpp
 
 !ios {
     HEADERS += src/daemon/DaemonManager.h
@@ -420,8 +422,6 @@ DISTFILES += \
     components/MobileHeader.qml \
     pages/intense.qml \
     IntenseConfig.js \
-    IntenseConfig.js \
-    IntenseConfig.js
     components/MobileHeader.qml
